backend/requirements.txt
fastapi
uvicorn[standard]
numpy
faster-whisper
pydantic
python-multipart
paho-mqtt